requirements.txt
fastapi 
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic-settings
python-dotenv
